Allow matching on WM_CLASS name
The X class hint contains two values: The name and the class.  For
certain applications such as wine, matching on the class is not useful
because it is too generic.  For cases like this, matching on the name
is needed.

This commit is an easy fix to match on window names; by specifying an
application name like "windowname.Foo" in the config file, this will
match all windows named "Foo".

Note that this is the WM_CLASS name property (which is usually static)
and not the WM_NAME property (which often changes during the run of a
program).