Iterate over X11 CRTCs directly
Currently we iterate over all outputs and then again over all CRTCs of
these outputs.

However, there is a chance that we pick up disabled outputs, which will
have a CRTC value of 0.
Calling XRRGetCrtcInfo with 0 will fail and make the entire program
crash. :(

If we instead iterate over all the CRTCs directly, we can be sure that
we do not call XRRGetCrtcInfo for an invalid CRTC value.

Fixes #341.