Use "Keysym" instead of "Keycode" in documentation
In X terminology, a keycode is a numerical identifier for an actual key
on the keyboard, which is mapped to a more expressive "keysym" by X
(e.g. see the `xmodmap` manpage).

The values specified in the configuration are actually keysyms, not
keycodes, so name them as such in the documentation.