Search xdg base directories for config file
This change makes touchegg more portable for users and systems that
follow the xdg base directory specifications but may not adhere to the
default directories such as `$HOME/.config` or `/etc/xdg`

For the user, a change has been made to `Paths::getHomePath` to look for
the path `$XDG_CONFIG_HOME` instead of `$HOME/.config`, creating the
directory `$HOME/.config` if an `$XDG_CONFIG_HOME` does not already
exist. Existing users should not notice a difference as the original
implementation also created the directory `$HOME/.config` if it did not
already exist.

For system-wide configuration, a change has been made to
`Paths::getSystemConfigFilePath` that first checks the paths in the
environment variable `$XDG_CONFIG_DIRS` for
`touchegg/touchegg.conf`. If it does not find a touchegg configuration in
the paths declared with `$XDG_CONFIG_DIRS` then it defaults to the
path declared in the compile time flag `SYSTEM_CONFIG_FILE_PATH` which
is usually /usr/share/.

So the order of looking for touchegg configuration is now

`$XDG_CONFIG_HOME/touchegg/touchegg.conf` (default: `$HOME/.config`)

-> `$XDG_CONFIG_DIRS/touchegg/touchegg.conf` (default: `/etc`, `etc/xdg`
for most distros)

-> `/usr/share/touchegg/touchegg.conf` (default for touchegg)

In addition small changes were made to the other path functions to
reflect now calling the xdg environment variables directly instead of
calling the defaults where they *should* be.

This should resolve issue #477